Extract attributes editor.
